all message validations moved to messaging folder
